Use webpack IgnorePlugin instead of externals
The documentation describes that optional dependencies should be marked as externals in webpack. This means that they will still show up in the source map.

The new setup simply ignores them, which is more correct. This was already reflected by Enzyme's own karma webpack configuration.

The React 0.14 ignore list was also written as a webpack config for consistency.