Test ReactWrapper simulate's event mocking
Based on enzyme's doc we should be able to use simulate()'s second
argument this way:

simulate(event [, mock])
- mock (Object [optional]): A mock event object that will be merged with
the event object passed to the handlers.

This test demonstrates a problem trying to simulate a change event on
an input field: we are unable to override the `target` property.

The same goes for the `currentTarget` property.