[Docs] Add link to new Preact adapter in README
Replace the link to the Preact adapter with the newer one from the
Preact org on GitHub.

This adapter supports all three Enzyme rendering modes as well as the upcoming
Preact v10 release (which is a rewrite of the internals).

It also avoids an indirect dependency on React and the need to install the
React-compatibility layer in order to use it.